test(expression): move lambda regression to it tests
