+++ src/query/expression/src/evaluator.rs
@@ -2098,73 +2098,3 @@ impl CheckStrictValue {
         DecimalDataType::from_value(value).unwrap().0.is_strict()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use databend_common_column::bitmap::Bitmap;
-
-    use super::Evaluator;
-    use crate::DataBlock;
-    use crate::FromData;
-    use crate::FunctionContext;
-    use crate::FunctionRegistry;
-    use crate::RemoteExpr;
-    use crate::ScalarRef;
-    use crate::types::DataType;
-    use crate::types::NullableType;
-    use crate::types::NumberDataType;
-    use crate::types::array::ArrayColumn;
-    use crate::types::nullable::NullableColumn;
-    use crate::types::number::Int64Type;
-    use crate::values::Column;
-    use crate::values::Value;
-
-    #[test]
-    fn test_run_lambda_ignores_physical_nullable_wrapper_for_non_nullable_array() {
-        let block = DataBlock::empty();
-        let func_ctx = FunctionContext::default();
-        let fn_registry = FunctionRegistry::empty();
-        let evaluator = Evaluator::new(&block, &func_ctx, &fn_registry);
-
-        let inner_ty = DataType::Nullable(Box::new(DataType::Number(NumberDataType::Int64)));
-        let array_ty = DataType::Array(Box::new(inner_ty.clone()));
-        let inner_column = NullableColumn::new_column(
-            Int64Type::from_data(vec![1i64, 2, 3]),
-            Bitmap::new_constant(true, 3),
-        );
-        let array_column = Column::Array(Box::new(ArrayColumn::new(
-            inner_column,
-            vec![0_u64, 3].into(),
-        )));
-
-        // Arrays with nullable elements may arrive wrapped physically, even when the
-        // logical argument type is not nullable.
-        let arg = Value::Column(NullableColumn::new_column(
-            array_column,
-            Bitmap::new_constant(true, 1),
-        ));
-        let lambda_expr = RemoteExpr::ColumnRef {
-            span: None,
-            id: 0,
-            data_type: inner_ty,
-            display_name: "x".to_string(),
-        };
-
-        let result = evaluator
-            .run_lambda(
-                "array_transform",
-                vec![arg],
-                vec![array_ty.clone()],
-                &lambda_expr,
-                &array_ty,
-            )
-            .unwrap();
-
-        let column = result.into_column().unwrap();
-        assert_eq!(column.data_type(), array_ty);
-        assert!(matches!(
-            unsafe { column.index_unchecked(0) },
-            ScalarRef::Array(_)
-        ));
-    }
-}
